requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pytesseract==0.3.10
pdfplumber==0.10.3
PyPDF2==3.0.1
Pillow==10.0.0
cryptography==41.0.3
bcrypt==4.0.1
faker==19.6.1
transformers==4.33.0
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
faiss-cpu==1.7.4
requests==2.31.0
redis==5.0.0
celery==5.3.1
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-asyncio==0.21.1
